Initial stab and a generic gabbi plugin for tempest
This is entirely the result of @sileht, who got things working for
gnocchi.

The end game here is a generic way to have a suite of YAML files
sitting around somewhere and run them, via tempest, and for it
to just work. This is currently some distance away from that
(because I got distracted finding nova bugs).