Extract a generic gabbi class
Subclass of the generic class determine the service_type and
service_name to be tested. The service_type drives the location of
the gabbits. Note that nothing prevents loading gabbits from
somewhere else and having multiple service types at play, this
change is just a way to clarify options.

The compute tests are moved into their own directory and some
glance/image tests are added.