Initial pass at allowing multiple services
Previously, the host was the endpoint of the service being tested.
This is limiting if what's really desired is testing multiple
services. So now we move towards setting environmen variable that
are the endpoints for any of the services that are going to be used.

Then these can be made available from any test.

Right now it is not like that. Right now it builds environments
based on the requested types. Better to just set them all and then
allow the host either be a primary service type endpoint, or a stub
if it really is going to be lots.

Committing now because tests are mostly passing (except for the
phantom failures). Cleanup to come.