Prove out multi-endpoint testing
multi/base.yaml looks at both the compute and placement services,
confirming that the one hypervisor that exists has a resource
provider with inventory with a proper allocation ratio for VCPU.

It turns out that each of the <TYPE>_SERVICE ENVIRON variables also
needs a <TYPE>_BASE otherwise it can be hard to reconstruct URLs
using links in respones. Sometimes they are full hosts. Sometimes
they are absolute URLs. Neither of this can be appended to a fully
qualified endpoint.