Establish a basic create/delete server arc
Move around some ENVIRON so that we always have the expected
values in the tests (it's cheap so why not?).

multi/base.yaml tries to take advantage of YAML anchors to avoid
repeating some of the longer substitutions[1] but I'm not sure if it
is going to be the right thing.

Next: do a resize.

[1] https://gabbi.readthedocs.io/en/latest/format.html#substitution