Use a GABBI_TEMPEST_PATH to locate gabbits
In this way the plugin becomes generic: point it at some directories
that contain gabbi YAML files that use the provided environment
variables (see the README.rst), what them run.

One concern that will need to be address is dealing with parallelism:
tempest doesn't make it easy to manipulate the group_regex.